+++ docs/release-notes.md
@@ -6,6 +6,7 @@
 * For the web UI:
   * If a JSON validator upload can't be saved to Pastebin (e.g. due to rate limits), it's now uploaded to Amazon S3 instead. Files uploaded to S3 expire after one month.
   * Updated the JSON validator for Content Patcher 1.10.0.
+  * Fixed JSON validator no longer letting you change format when viewing results.
 
 ## 3.0.1
 Released 02 December 2019 for Stardew Valley 1.4.0.1.